Removed ancient script_macros_common.hpp, replaced with external include to CBA
