+++ A3A/addons/core/Includes/script_macros.hpp
@@ -1,4 +1,4 @@
-#include "script_macros_common.hpp"
+#include "\x\cba\addons\main\script_macros_common.hpp"
 
 #undef PREP
 #undef PREPSUB
@@ -6,7 +6,23 @@
 #define PREPSUB(folder,fncName) FUNC(fncName) = compile preprocessFileLineNumbers QPATHTOF(folder\DOUBLES(fn,fncName).sqf)
 
 #undef VARDEF
-#define VARDEF(Var, Def) (if (isNil #Var) then {Def} else {Var})
+#define VARDEF RETDEF
 
-#define ADDONLOADED(addon) EADDONLOADED(A3A,addon)
-#define EADDONLOADED(prefix,addon)(isClass (configFile/QUOTE(CfgPatches)/QDOUBLES(prefix,addon)))
+/* -------------------------------------------
+Macro: XOR
+    Evaluates to true if exactly one of both values is true
+
+Parameters:
+    VAR1 - the first variable to evaluate
+    VAR2 - the second variable to evaluate
+
+Example:
+    (begin example)
+        // return "true" if exactly one of a and b is true
+        XOR(a, b);
+    (end)
+
+Author:
+    Bohemia Interactive (https://community.bistudio.com/wiki/Operators)
+------------------------------------------- */
+#define XOR(VAR1,VAR2) (((VAR1) || (VAR2)) && !((VAR1) && (VAR2)))
